fix: process handing
Fix the handling of the CodeNarc process when it's not accessible via
its web API.

We now maintain a handle to the process so if we can't get a response to
the initial ping requests we can still shut it down.

In addition to this correctly handle the timeout in the error case.