fix: UnnecessaryGString for strings with slashes
Fix the processing of UnnecessaryGString fixes when the string in
question includes one or more backslashes.

This is caused by CodeNarc/CodeNarc#749 which results in the output
not escaping backslashes in the message we use to determine the
value of the string.

This will need to be reverted when we update to a fixed CodeNar version.

Fixes #299