chore: update node to 18 lts (#310)
* chore: update node to 18 lts

Update to the node 18 lts so we're on a supported version.

* ci: enable debug on test workflows

Enable debug on all test workflows so we can understand any failures.

* chore: reduce debug log spam

Reduce the amount of output in the log when running in debug.

* feat: CodeNarcServer logging

Add slf4j logging to CodeNarcServer so that we can control the log
level.

* chore: add trace level logging

Add support for trace level debugging so we can reduce the spam while
investigating issues for the log entries which are particularly large.

Trace level logging can be enabled by add npm-groovy-lint-trace to the
DEBUG environment variable.