ci: predictable server builds and port option
Implement the ability to create predictable server builds so that the
workflows can validate that it's been updated as needed.

This adds the  workflow lint which validates that everything has been
built and lint fixes applied.

This switches running the server to use -jar to reduce complexity now
the manifest is correctly created.

Add missing cli jars not needed for source runs, causing them to be
missed in previous commit.

Moved scripts into their own folder to keep the root directory cleaner.

This includes a lint:fix and build changes, unrelated to this as they
hadn't been run in the past.

Removed try catch in server runCodeNarc as it was resulting in
exceptions being logged twice.