fix: missing parse errors with --noserver
Refactor request processing so that the flow is the same with and
without --noserver command line option which was preventing parse errors
being reported when --noserver was specified.

Add duration in milliseconds to request logging so we easily monitor how
long the server side requests are taking.

Explicitly handle 404 errors so we don't needlessly retry them with a
slow direct request.

Use chaining to improve readability.

Switch to java.time.* instead of Date as recommended by CodeNarc.

Fixes #323