Adds GitHub Dependents Info workflow
Adds a new workflow to automatically update markdown files with GitHub dependents information.

The workflow runs on pushes to the main, master, and dependent-actions branches, as well as on a weekly schedule. It uses the `nvuillam/github-dependents-info` action to generate the dependents information and creates a pull request to update the markdown files.